RAI: Add session management
Begin to create more generic session managment.
Integrate the old access management into a session class.

NOTE: This and subsequent commits affecting PHP code in /rai
may also contain whitespace edits (of significant size) along with
code optimizations and/or changes that impove general readability.

These changes may or may not be mentioned in the commit messages.
Due to the amount of re-formatting, It's also possible that
a commit in the series may be broken. It's not intended that
the user should checkout a commit in this series and expect
everything to work.
The commit that ends this series of edits will contain a note
to that effect.