Fix for names in databases.
When adding a name to an osmo-nitb hlr, use unidecode() on it first
as that vty command won't accept accents.

At the same time, let's NOT strip the accents from the name before
adding to postrges database, a change which was added way back in
d62398a1a1b9cdac3591fd5440c01b6f3e712fb8 but no reason was given in
that commit.